lab_8

+++ flaskProject/blueprint.py
@@ -1,15 +1,30 @@
 import sqlalchemy
 from flask import Blueprint, request, jsonify
 import marshmallow
+from flask_bcrypt import check_password_hash
+from flask_jwt_extended import jwt_required, create_access_token, get_jwt_identity, get_jwt
+import sys
+from datetime import datetime
+from datetime import timedelta
+from datetime import timezone
+sys.path.append('../LB-5/Application_Programming_Project')
 import db_utils
 from schemas import *
 from models import *
-
+from config import jwt
 api_blueprint = Blueprint('api', __name__)
 StudentID = 1
+session = Session()
 
 errors = Blueprint('errors', __name__)
 
+
+@jwt.token_in_blocklist_loader
+def check_if_token_revoked(jwt_header, jwt_payload: dict) -> bool:
+    jti = jwt_payload["jti"]
+    token = session.query(TokenBlocklist.id).filter_by(jti=jti).scalar()
+
+    return token is not None
 
 @errors.app_errorhandler(sqlalchemy.exc.NoResultFound)
 def handle_error(error):
@@ -59,15 +74,16 @@ def hello_world_ex():
 def hello_world():
     return f'Hello World {StudentID}', 200
 
-
+#checked
 @api_blueprint.route("/user", methods=["POST"])
 def create_user():
     user_data = UserCreate().load(request.json)
     user = db_utils.create_entry(Users, **user_data)
     return jsonify(UserInfo().dump(user))
 
-
+#checked
 @api_blueprint.route("/user/<int:user_id>", methods=["GET"])
+@jwt_required()
 def get_user_by_id(user_id):
     try:
         user = db_utils.get_entry_by_id(Users, user_id)
@@ -85,8 +101,88 @@ def get_user_by_id(user_id):
     return jsonify(UserInfo().dump(user))
 
 
+#checked
 @api_blueprint.route("/user/<int:user_id>", methods=["PUT"])
+@jwt_required()
 def update_user(user_id):
+    try:
+        user = db_utils.get_entry_by_id(Users, user_id)
+    except sqlalchemy.exc.NoResultFound:
+        response = {
+            'error': {
+                'code': 404,
+                'type': 'NOT_FOUND',
+                'message': 'User not found'
+            }
+        }
+
+        return jsonify(response), 404
+    if user_id != get_jwt_identity():
+        return jsonify({'error': "Access denied"}), 403   
+
+    user_data = UserUpdate().load(request.json)
+    user_updated = db_utils.update_entry(Users, user_id, **user_data)
+    return jsonify(UserInfo().dump(user_updated))
+
+#checked
+@api_blueprint.route("/user", methods=["DELETE"])
+@jwt_required()
+def delete_user():
+    try:
+        db_utils.get_entry_by_id(Users, get_jwt_identity())
+    except sqlalchemy.exc.NoResultFound:
+        response = {
+            'error': {
+                'code': 404,
+                'type': 'NOT_FOUND',
+                'message': 'User not found'
+            }
+        }
+
+        return jsonify(response), 404
+
+    db_utils.delete_entry(Users, get_jwt_identity())
+    logout_user()
+    return jsonify({"code": 200, "message": "OK", "type": "OK"})
+
+
+#checked
+@api_blueprint.route("/user/login", methods=["GET"])
+def login_user():
+    userLog = LoginUser().load(request.get_json())
+    sys_user = session.query(Users).filter_by(email=userLog['email']).first()
+    if not sys_user:
+        response = {
+            'error': {
+                'code': 404,
+                'type': 'NOT_FOUND',
+                'message': 'User not found'
+            }
+        }
+        return jsonify(response), 404
+
+    if check_password_hash(sys_user.password, userLog['password']):
+            return jsonify(access_token=create_access_token(identity=sys_user.id)), 200
+    else:
+        return jsonify({"Error": "Wrong password"}), 401
+    
+
+#checked
+@api_blueprint.route("/user/logout", methods=["DELETE"])
+@jwt_required()
+def logout_user():
+    jti = get_jwt()["jti"]
+    now = datetime.now(timezone.utc)
+    session.add(TokenBlocklist(jti=jti, created_at=now))
+    session.commit()
+    return jsonify(msg="JWT revoked")
+
+#checked
+@api_blueprint.route("/wallet", methods=["POST"])
+@jwt_required()
+def create_wallet():
+    user_id = get_jwt_identity()
+    wallet_funds = WalletCreate().load(request.json)
     try:
         db_utils.get_entry_by_id(Users, user_id)
     except sqlalchemy.exc.NoResultFound:
@@ -100,57 +196,7 @@ def update_user(user_id):
 
         return jsonify(response), 404
 
-    user_data = UserUpdate().load(request.json)
-    user_updated = db_utils.update_entry(Users, user_id, **user_data)
-    return jsonify(UserInfo().dump(user_updated))
-
-
-@api_blueprint.route("/user/<int:user_id>", methods=["DELETE"])
-def delete_user(user_id):
-    try:
-        user = db_utils.get_entry_by_id(Users, user_id)
-    except sqlalchemy.exc.NoResultFound:
-        response = {
-            'error': {
-                'code': 404,
-                'type': 'NOT_FOUND',
-                'message': 'User not found'
-            }
-        }
-
-        return jsonify(response), 404
-
-    db_utils.delete_entry(Users, user_id)
-    return jsonify({"code": 200, "message": "OK", "type": "OK"})
-
-
-@api_blueprint.route("/user/login", methods=["GET"])
-def login_user():
-    return jsonify({"code": 200, "message": "NOT IMPLEMENTED", "type": "OK"})
-
-
-@api_blueprint.route("/user/logout", methods=["GET"])
-def logout_user():
-    return jsonify({"code": 200, "message": "NOT IMPLEMENTED", "type": "OK"})
-
-
-@api_blueprint.route("/wallet", methods=["POST"])
-def create_wallet():
-    wallet_data = WalletCreate().load(request.json)
-    try:
-        db_utils.get_entry_by_id(Users, wallet_data["user_id"])
-    except sqlalchemy.exc.NoResultFound:
-        response = {
-            'error': {
-                'code': 404,
-                'type': 'NOT_FOUND',
-                'message': 'User not found'
-            }
-        }
-
-        return jsonify(response), 404
-
-    if wallet_data["funds"] < 0:
+    if wallet_funds["funds"] < 0:
         response = {
             'error': {
                 'code': 400,
@@ -160,19 +206,24 @@ def create_wallet():
         }
 
         return jsonify(response), 400
-
+    wallet_data = {
+        'funds': wallet_funds['funds'],
+        'user_id': user_id
+    }
     wallet = db_utils.create_entry(Wallets, **wallet_data)
     return jsonify(WalletInfo().dump(wallet))
 
-
+#checked
 @api_blueprint.route("/wallet", methods=["GET"])
+@jwt_required()
 def get_user_wallets():
-    user_id = 4
+    user_id = get_jwt_identity()
     wallets = db_utils.get_wallets_by_user_id(user_id)
     return jsonify(WalletInfo().dump(wallets, many=True))
 
-
+#checked
 @api_blueprint.route("/wallet/<int:wallet_id>", methods=["GET"])
+@jwt_required()
 def get_wallet_by_id(wallet_id):
     try:
         wallet = db_utils.get_entry_by_id(Wallets, wallet_id)
@@ -187,13 +238,17 @@ def get_wallet_by_id(wallet_id):
 
         return jsonify(response), 404
 
+    if wallet.user_id != get_jwt_identity():
+        return jsonify({'error': "Access denied"}), 403
+
     return jsonify(WalletInfo().dump(wallet))
 
-
+#checked
 @api_blueprint.route("/wallet/<int:wallet_id>", methods=["PUT"])
+@jwt_required()
 def update_wallet(wallet_id):
     try:
-        db_utils.get_entry_by_id(Wallets, wallet_id)
+        wallet = db_utils.get_entry_by_id(Wallets, wallet_id)
     except sqlalchemy.exc.NoResultFound:
         response = {
             'error': {
@@ -204,16 +259,20 @@ def update_wallet(wallet_id):
         }
 
         return jsonify(response), 404
+
+    if wallet.user_id != get_jwt_identity():
+        return jsonify({'error': "Access denied"}), 403
 
     wallet_data = WalletUpdate().load(request.json)
     wallet_updated = db_utils.update_entry(Wallets, wallet_id, **wallet_data)
     return jsonify(WalletInfo().dump(wallet_updated))
 
-
+#checked
 @api_blueprint.route("/wallet/<int:wallet_id>", methods=["DELETE"])
+@jwt_required()
 def delete_wallet(wallet_id):
     try:
-        db_utils.get_entry_by_id(Wallets, wallet_id)
+        wallet = db_utils.get_entry_by_id(Wallets, wallet_id)
     except sqlalchemy.exc.NoResultFound:
         response = {
             'error': {
@@ -224,12 +283,16 @@ def delete_wallet(wallet_id):
         }
 
         return jsonify(response), 404
+
+    if wallet.user_id != get_jwt_identity():
+        return jsonify({'error': "Access denied"}), 403
 
     db_utils.delete_entry(Wallets, wallet_id)
     return jsonify({"code": 200, "message": "OK", "type": "OK"})
 
 
 @api_blueprint.route("/wallet/make-transfer", methods=["POST"])
+@jwt_required()
 def wallet_make_transfer():
     transfer_data = TransferCreate().load(request.json)
 
@@ -237,7 +300,7 @@ def wallet_make_transfer():
     receiver = transfer_data["to_wallet_id"]
 
     try:
-        db_utils.get_entry_by_id(Wallets, sender)
+        wallet = db_utils.get_entry_by_id(Wallets, sender)
     except sqlalchemy.exc.NoResultFound:
         response = {
             'error': {
@@ -262,6 +325,9 @@ def wallet_make_transfer():
 
         return jsonify(response), 404
 
+    if wallet.user_id != get_jwt_identity():
+        return jsonify({'error': "You dont have access to this wallet"}), 403
+
     if db_utils.get_entry_by_id(Wallets, sender).funds < transfer_data["amount"]:
         response = {
             'error': {
@@ -282,10 +348,11 @@ def wallet_make_transfer():
 
 
 @api_blueprint.route("/wallet/<int:wallet_id>/transfers", methods=["GET"])
+@jwt_required()
 def get_transfers(wallet_id):
 
     try:
-        db_utils.get_entry_by_id(Wallets, wallet_id)
+        user_wallet = db_utils.get_entry_by_id(Wallets, wallet_id)
     except sqlalchemy.exc.NoResultFound:
         response = {
             'error': {
@@ -296,6 +363,9 @@ def get_transfers(wallet_id):
         }
 
         return jsonify(response), 404
+
+    if user_wallet.user_id != get_jwt_identity():
+        return jsonify({'error': "Access denied"}), 403
 
     transfers = db_utils.get_transfers_by_wallet_id(wallet_id)
     return jsonify(TransferInfo().dump(transfers, many=True))
